# design/openapi.json
{"openapi": "3.0.2", "info": {"title": "Orbis2 Backend", "version": "0.0.2"}, "paths": {"/getAnnotations": {"get": {"summary": "Get Annotations", "operationId": "get_annotations_getAnnotations_get", "parameters": [{"required": false, "schema": {"title": "Run Id", "type": "integer"}, "name": "run_id", "in": "query"}, {"required": false, "schema": {"title": "Document Id", "type": "integer"}, "name": "document_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/getDocuments": {"get": {"summary": "Get Documents", "operationId": "get_documents_getDocuments_get", "parameters": [{"required": false, "schema": {"title": "Run Id", "type": "integer"}, "name": "run_id", "in": "query"}, {"required": false, "schema": {"title": "Corpus Id", "type": "integer"}, "name": "corpus_id", "in": "query"}, {"required": false, "schema": {"title": "Page Size", "type": "integer"}, "name": "page_size", "in": "query"}, {"required": false, "schema": {"title": "Skip", "type": "integer", "default": 0}, "name": "skip", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/getDocument": {"get": {"summary": "Get Document", "operationId": "get_document_getDocument_get", "parameters": [{"required": true, "schema": {"title": "Document Id", "type": "integer"}, "name": "document_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/getRuns": {"get": {"summary": "Get Runs", "operationId": "get_runs_getRuns_get", "parameters": [{"required": false, "schema": {"title": "Corpus Id", "type": "integer"}, "name": "corpus_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/getCorpora": {"get": {"summary": "Get Corpora", "operationId": "get_corpora_getCorpora_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/getCorpus": {"get": {"summary": "Get Corpus", "operationId": "get_corpus_getCorpus_get", "parameters": [{"required": true, "schema": {"title": "Corpus Id", "type": "integer"}, "name": "corpus_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/addCorpus": {"post": {"summary": "Add Corpus", "operationId": "add_corpus_addCorpus_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_add_corpus_addCorpus_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/addRun": {"post": {"summary": "Add Run", "operationId": "add_run_addRun_post", "parameters": [{"required": true, "schema": {"title": "Run Name", "type": "string"}, "name": "run_name", "in": "query"}, {"required": true, "schema": {"title": "Run Description", "type": "string"}, "name": "run_description", "in": "query"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Corpus"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/deleteRun": {"delete": {"summary": "Delete Run", "operationId": "delete_run_deleteRun_delete", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Run"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/addAnnotation": {"post": {"summary": "Add Annotation", "operationId": "add_annotation_addAnnotation_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Annotation"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/deleteAnnotationFromDocument": {"delete": {"summary": "Delete Annotation From Document", "operationId": "delete_annotation_from_document_deleteAnnotationFromDocument_delete", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Annotation"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/deleteCorpus": {"delete": {"summary": "Delete Corpus", "operationId": "delete_corpus_deleteCorpus_delete", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Corpus"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/nextDocument": {"get": {"summary": "Next Document", "operationId": "next_document_nextDocument_get", "parameters": [{"required": true, "schema": {"title": "Run Id", "type": "integer"}, "name": "run_id", "in": "query"}, {"required": true, "schema": {"title": "Document Id", "type": "integer"}, "name": "document_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/previousDocument": {"get": {"summary": "Previous Document", "operationId": "previous_document_previousDocument_get", "parameters": [{"required": true, "schema": {"title": "Run Id", "type": "integer"}, "name": "run_id", "in": "query"}, {"required": true, "schema": {"title": "Document Id", "type": "integer"}, "name": "document_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/countDocuments": {"get": {"summary": "Count Documents", "operationId": "count_documents_countDocuments_get", "parameters": [{"required": true, "schema": {"title": "Run Id", "type": "integer"}, "name": "run_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/colorPalettes": {"get": {"summary": "Color Palettes", "operationId": "color_palettes_colorPalettes_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/corpusAnnotationTypes": {"get": {"summary": "Get Corpus Annotation Types", "operationId": "get_corpus_annotation_types_corpusAnnotationTypes_get", "parameters": [{"required": true, "schema": {"title": "Corpus Id", "type": "integer"}, "name": "corpus_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Annotation": {"title": "Annotation", "required": ["key", "surface_forms", "start_indices", "end_indices", "annotation_type", "annotator"], "type": "object", "properties": {"key": {"title": "Key", "type": "string"}, "surface_forms": {"title": "Surface Forms", "anyOf": [{"type": "array", "items": {"type": "string"}}, {"type": "string"}]}, "start_indices": {"title": "Start Indices", "anyOf": [{"type": "array", "items": {"type": "integer"}}, {"type": "integer"}]}, "end_indices": {"title": "End Indices", "anyOf": [{"type": "array", "items": {"type": "integer"}}, {"type": "integer"}]}, "annotation_type": {"$ref": "#/components/schemas/AnnotationType"}, "annotator": {"$ref": "#/components/schemas/Annotator"}, "run_id": {"title": "Run Id", "type": "integer"}, "document_id": {"title": "Document Id", "type": "integer"}, "metadata": {"title": "Metadata", "type": "array", "items": {"$ref": "#/components/schemas/Metadata"}}, "timestamp": {"title": "Timestamp", "type": "string", "format": "date-time"}}}, "AnnotationType": {"title": "AnnotationType", "required": ["name"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "color_id": {"title": "Color Id", "type": "integer"}}}, "Annotator": {"title": "Annotator", "required": ["name", "roles"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "roles": {"title": "Roles", "type": "array", "items": {"$ref": "#/components/schemas/Role"}}, "password": {"title": "Password", "type": "string", "default": "02cc5d05"}}}, "Body_add_corpus_addCorpus_post": {"title": "Body_add_corpus_addCorpus_post", "required": ["corpus"], "type": "object", "properties": {"corpus": {"$ref": "#/components/schemas/Corpus"}, "documents": {"title": "Documents", "type": "array", "items": {"$ref": "#/components/schemas/Document"}}}}, "Corpus": {"title": "Corpus", "required": ["name", "supported_annotation_types"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "supported_annotation_types": {"title": "Supported Annotation Types", "type": "array", "items": {"$ref": "#/components/schemas/AnnotationType"}}}}, "Document": {"title": "Document", "required": ["content", "key"], "type": "object", "properties": {"content": {"title": "Content", "type": "string"}, "key": {"title": "Key", "type": "string"}, "run_id": {"title": "Run Id", "type": "integer"}, "metadata": {"title": "Metadata", "type": "array", "items": {"$ref": "#/components/schemas/Metadata"}}, "done": {"title": "Done", "type": "boolean", "default": false}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "Metadata": {"title": "Metadata", "required": ["key", "value"], "type": "object", "properties": {"key": {"title": "Key", "type": "string"}, "value": {"title": "Value", "type": "string"}}}, "Role": {"title": "Role", "required": ["name"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}}}, "Run": {"title": "Run", "required": ["name"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "description": {"title": "Description", "type": "string"}, "corpus": {"$ref": "#/components/schemas/Corpus"}, "document_annotations": {"title": "Document Annotations", "type": "object", "additionalProperties": {"type": "array", "items": {"$ref": "#/components/schemas/Annotation"}}}, "parents": {"title": "Parents", "type": "array", "items": {"$ref": "#/components/schemas/Run"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}